add Task2-3 solution and tests

+++ src/main/java/edu/hw7/Task1.java
@@ -2,19 +2,19 @@ package edu.hw7;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class Task1 {
     private Task1() {
     }
 
-    public static int multiThreadIncrement(int threadsCount, int incrementCount) {
-        AtomicInteger atomic = new AtomicInteger();
+    public static long multiThreadIncrement(int threadsCount, long incrementCount) {
+        AtomicLong atomic = new AtomicLong();
         List<Thread> threads = new ArrayList<>();
 
         for (int i = 0; i < threadsCount; i++) {
             threads.add(new Thread(() -> {
-                for (int j = 0; j < incrementCount; j++) {
+                for (long j = 0; j < incrementCount; j++) {
                     atomic.incrementAndGet();
                 }
             }));
